Fix test passing.
